Revert "Renaming to QueueDAG"
This reverts commit 800701bcdbcd3299ae1631cbed64ac77dd1354cd.
