+++ webber/__init__.py
@@ -8,7 +8,7 @@ __version__ = '0.1.2'
 __supported__ = ("linux", "linux2", "win32")
 
 if _sys.platform not in __supported__:
-    _logging.warning(f"Webber {__version__} is only supported on these platforms: {', '.join(__supported__)}")
+    _logging.warn(f"Webber {__version__} is only supported on these platforms: {', '.join(__supported__)}")
 
-from .core import DAG, Condition, QueueDAG
+from .core import DAG, Condition
 from .xcoms import Promise
