+++ webber/etl/complex_dag.py
@@ -1,4 +1,5 @@
-from webber import Promise, DAG, QueueDAG
+import webber.etl as etl
+from webber import Promise, DAG
 import time
 
 def writer(x):
@@ -10,7 +11,7 @@ def writer(x):
 main_dag = DAG()
 name = main_dag.add_node(lambda: "World")
 
-queue_dag = QueueDAG()
+queue_dag = etl.AsyncDAG()
 greeting = queue_dag.add_node(lambda n: f"Hello, {n}", Promise(name), iterator=200)
 writing  = queue_dag.add_node(writer, Promise(greeting))
 queue_dag.add_edge(greeting, writing)
@@ -19,8 +20,6 @@ queue_dag.add_edge(greeting, writing)
 main_dag.add_node(queue_dag.execute, name, Promise(name), print_exc = True)
 main_dag.add_edge(name, queue_dag.execute)
 
-# Nesting DAGs leads to expectedly slower behavior.
-# TODO: Should Nested DAGs be split off as separate processes to receive adequate resources?
 t = time.time()
 main_dag.execute()
 print(time.time() - t)
@@ -33,3 +32,4 @@ t2 = time.time()
 for i in range(200):
     writer("Hello, World")
 print(time.time() - t2)
+
