+++ webber/etl/testing.py
@@ -1,5 +1,6 @@
-import webber
+import webber.etl as etl
 import time
+from webber import Promise
 
 def writer(x):
     fd = open('myfile.txt', 'a+')
@@ -7,9 +8,9 @@ def writer(x):
     fd.close()
 
 
-dag = webber.QueueDAG()
+dag = etl.AsyncDAG()
 x = dag.add_node(lambda: "1", iterator=100)
-y = dag.add_node(writer, webber.Promise(x))
+y = dag.add_node(writer, Promise(x))
 dag.add_edge(x, y)
 t = time.time()
 dag.execute(print_exc=True)
